Update Complete 14 bus NR
Dear author,
First of all, thank you for this code and for sharing this with the community.
Running the program for a four bus bar system and comparing the answer with a comercial program, I observed that the results of J2 and J4 elements is not correct. 
I kindly ask you to revise the formula as I did in the comments.
Regards.